React_web_app/server/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
pandas==0.25.3
google-cloud-bigquery==1.23.0
